micro: extract chat pane header
