+++ frontend/src/ui/index.ts
@@ -97,6 +97,7 @@ export { ModelLogo } from "./model-logo";
 export { HuggingFaceModelCardModal } from "./huggingface-model-card";
 export { AgentModelPicker } from "./agent-model-picker";
 export { AgentQueuePanel } from "./agent-queue-panel";
+export { AgentChatPaneHeader } from "./agent-chat-pane-header";
 export { AgentLoadedContextTabs, AgentMentionPicker } from "./agent-composer-context";
 export type { FileMentionRow, MentionRow } from "./agent-composer-context";
